Build 0044
No difference except the seeds position

+++ NORDARK/Assets/Scripts/Node.cs
@@ -21,8 +21,8 @@ public class Node : ScriptableObject
     public float[] LeastCostList; //List of costs for different POIs for risk calculation
     public float riskFactor; //riskFactor calculation happens in Graph.cs
 
-    public float x_index;
-    public float z_index;
+    public int x_index;
+    public int z_index;
 
     public Vector3 globalposition;
